// src/generated/api/iam/iam.schemas.ts
export type UserDtoUserType = (typeof UserDtoUserType)[keyof typeof UserDtoUserType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserDtoUserType = {
  STAFF: 'STAFF',
  SYSTEM: 'SYSTEM',
} as const;

export type UserDtoStatus = (typeof UserDtoStatus)[keyof typeof UserDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserDtoStatus = {
  ACTIVE: 'ACTIVE',
  LOCKED: 'LOCKED',
  INACTIVE: 'INACTIVE',
} as const;

export interface UserDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  displayName: string;
  maskedEmail: string;
  userType: UserDtoUserType;
  status: UserDtoStatus;
  permissionVersion: number;
  /** @minimum 0 */
  failedLoginAttempts: number;
  mustChangePassword: boolean;
  lockedAt?: string;
  lockReason?: string;
  assignments: UserRoleAssignmentDto[];
}
// ...
export type UserRoleAssignmentDtoStatus =
  (typeof UserRoleAssignmentDtoStatus)[keyof typeof UserRoleAssignmentDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRoleAssignmentDtoStatus = {
  ACTIVE: 'ACTIVE',
} as const;

export interface UserRoleAssignmentDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  /** @pattern ^[1-9][0-9]*$ */
  userId: string;
  /** @pattern ^[1-9][0-9]*$ */
  roleId: string;
  roleCode: string;
  scopeType: ScopeType;
  /** @pattern ^[1-9][0-9]*$ */
  branchId?: string;
  status: UserRoleAssignmentDtoStatus;
  validFrom: string;
}
// ...
export type AssignableStaffRoleCode =
  (typeof AssignableStaffRoleCode)[keyof typeof AssignableStaffRoleCode];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AssignableStaffRoleCode = {
  BRANCH_MANAGER: 'BRANCH_MANAGER',
  STAFF: 'STAFF',
} as const;

export type ScopeType = (typeof ScopeType)[keyof typeof ScopeType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ScopeType = {
  GLOBAL: 'GLOBAL',
  BRANCH: 'BRANCH',
} as const;